fix(step3): separate Racing class with RacingCarController
Racing 클래스에서 차량을 만들어주고, 시행횟수를 지정하는것이 역할이 너무 많아진다는 느낌을 받았습니다.
따라서 Racing에서의 많은 역할을 어떻게 분리하지 생각을 하다가 RacingCarController를 만들어서 Controller에 역할을 분리해뒀습니다.
