+++ RacingCar/Racing.swift
@@ -9,38 +9,15 @@ import Foundation
 
 final class Racing {
     
-    // MARK: - Property
-    private(set) var cars: [Car] = []
-    private(set) var tryCount: Int = 0
-    
     // MARK: - Method
-    func start(carCount: Int, tryCount: Int) throws {
-        createCars(input: carCount)
-        self.tryCount = tryCount
-        try racing()
-    }
-    
-    private func createCars(input: Int) {
-        var count = input
-        while count != 0 {
-            cars.append(Car())
-            count -= 1
-        }
-    }
-    
-    private func racing() throws {
-        try runRacing {
-            RacingCarInputOutput.printResult(cars)
-        }
-    }
-    
-    func runRacing(completion: () -> Void) throws {
+    func runRacing(tryCount: Int, cars: [Car], completion: ([Car]) -> Void) throws {
+        var tryCount = tryCount
         while tryCount != 0 {
             try cars.forEach { car in
                 try car.move(randomNumber: car.generateRandomNumber())
             }
             tryCount -= 1
-            completion()
+            completion(cars)
         }
     }
 }
