+++ RacingCar/main.swift
@@ -7,7 +7,7 @@
 import Foundation
 
 do {
-    try Racing()
+    try RacingCarController()
         .start(carCount: try RacingCarInputOutput.inputCarCount(),
                tryCount: try RacingCarInputOutput.inputTryCount())
 } catch {
