+++ RacingCarTest/RacingCarTests/RacingTests.swift
@@ -10,6 +10,10 @@ import XCTest
 class RacingTests: XCTestCase {
     
     func test_레이싱_처음부터_잘_돌아가는지_테스트() throws {
-        XCTAssertNoThrow(try Racing().start(carCount: 4, tryCount: 4))
+        var cars: [Car] = []
+        for _ in 0 ..< 4 {
+            cars.append(Car())
+        }
+        XCTAssertNoThrow(try Racing().runRacing(tryCount: 3, cars: cars, completion: { _ in }))
     }
 }
